# app/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuickEx Backend API
  description: >
    QuickEx is a Stellar-based payment platform enabling payment links,
    recurring subscriptions, username resolution, marketplace trading,
    fiat on/off ramps, and webhook-driven event notifications.

    **Base URL (production):** `https://api.quickex.example.com`
    **Base URL (development):** `http://localhost:3000`

    Interactive Swagger UI is available at `/docs` when the server is running.
  version: 1.0.0
  contact:
    name: QuickEx API Support
    url: https://github.com/quickex/quickex
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.quickex.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - {}
paths:
...
  /links/recurring:
    post:
      tags: [recurring-payments]
      operationId: createRecurringLink
      summary: Create a recurring payment link
      description: Creates a subscription-style payment link with specified frequency.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecurringPaymentLinkRequest"
      responses:
        "201":
          description: Recurring link created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/RecurringPaymentLinkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      tags: [recurring-payments]
      operationId: listRecurringLinks
      summary: List recurring payment links
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RecurringStatus"
        - name: username
          in: query
          schema:
            type: string
        - name: destination
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: List of recurring payment links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRecurringPaymentsResponse"
  /links/recurring/{id}:
    get:
      tags: [recurring-payments]
      operationId: getRecurringLink
      summary: Get recurring payment link by ID
      parameters:
        - $ref: "#/components/parameters/RecurringLinkId"
      responses:
        "200":
          description: Recurring link details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/RecurringPaymentLinkResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [recurring-payments]
      operationId: updateRecurringLink
      summary: Update recurring payment link
      parameters:
        - $ref: "#/components/parameters/RecurringLinkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecurringPaymentLinkRequest"
      responses:
        "200":
          description: Link updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/RecurringPaymentLinkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/recurring/{id}/cancel:
    post:
      tags: [recurring-payments]
      operationId: cancelRecurringLink
      summary: Cancel recurring payment link
      parameters:
        - $ref: "#/components/parameters/RecurringLinkId"
      responses:
        "200":
          description: Link cancelled
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/recurring/{id}/pause:
    post:
      tags: [recurring-payments]
      operationId: pauseRecurringLink
      summary: Pause recurring payment link
      parameters:
        - $ref: "#/components/parameters/RecurringLinkId"
      responses:
        "200":
          description: Link paused
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/recurring/{id}/resume:
    post:
      tags: [recurring-payments]
      operationId: resumeRecurringLink
      summary: Resume recurring payment link
      parameters:
        - $ref: "#/components/parameters/RecurringLinkId"
      responses:
        "200":
          description: Link resumed
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /links/recurring/{id}/executions:
    get:
      tags: [recurring-payments]
      operationId: getExecutionHistory
      summary: Get execution history
      parameters:
        - $ref: "#/components/parameters/RecurringLinkId"
      responses:
        "200":
          description: Execution history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecurringPaymentExecution"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Transactions
  # ---------------------------------------------------------------------------
...
  /telegram/link/{telegramId}:
    delete:
      tags: [Telegram]
      operationId: unlinkTelegram
      summary: Unlink Telegram account
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Account unlinked
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Reconciliation
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    RecurringLinkId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Recurring payment link ID

  # -------------------------------------------------------------------------
  # Reusable Responses
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            request_id:
              type: string
              format: uuid
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            details:
              type: object

    # --- Health ---
...
    RecurringStatus:
      type: string
      enum: [active, paused, completed, cancelled]
    FrequencyType:
      type: string
      enum: [daily, weekly, monthly, yearly]
    CreateRecurringPaymentLinkRequest:
      type: object
      required: [amount, asset, frequency]
      properties:
        amount:
          type: number
          minimum: 1
        asset:
          type: string
        assetIssuer:
          type: string
        username:
          type: string
        destination:
          type: string
        payerPublicKey:
          type: string
          description: Optional Stellar public key for payer notification delivery.
        frequency:
          $ref: "#/components/schemas/FrequencyType"
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        totalPeriods:
          type: integer
          minimum: 1
        memo:
          type: string
        memoType:
          type: string
          enum: [text, id, hash, return]
        referenceId:
          type: string
        privacyEnabled:
          type: boolean
          default: false
    UpdateRecurringPaymentLinkRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
        frequency:
          $ref: "#/components/schemas/FrequencyType"
        endDate:
          type: string
          format: date-time
        totalPeriods:
          type: integer
          minimum: 1
        memo:
          type: string
        referenceId:
          type: string
    RecurringPaymentLinkResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        destination:
          type: string
        payerPublicKey:
          type: string
        amount:
          type: number
        asset:
          type: string
        assetIssuer:
          type: string
        frequency:
          $ref: "#/components/schemas/FrequencyType"
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        totalPeriods:
          type: integer
        executedCount:
          type: integer
        nextExecutionDate:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/RecurringStatus"
        memo:
          type: string
        memoType:
          type: string
        referenceId:
          type: string
        privacyEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RecurringPaymentExecution:
      type: object
      properties:
        id:
          type: string
        periodNumber:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        executedAt:
          type: string
          format: date-time
        amount:
          type: number
        asset:
          type: string
        status:
          type: string
          enum: [pending, processing, success, failed, skipped]
        transactionHash:
          type: string
        failureReason:
          type: string
        retryCount:
          type: integer
        notificationSent:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ListRecurringPaymentsResponse:
      type: object
      properties:
        success:
          type: boolean
        total:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/RecurringPaymentLinkResponse"
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        limit:
          type: integer

    # --- Transactions ---
...